rename diff files
