+++ packages/sdk/src/diff/allowances.test.ts
@@ -1,7 +1,7 @@
 import { expect, it, suite } from "vitest"
 import { hexlify, zeroPadValue } from "ethers"
 import { Allowance } from "zodiac-roles-deployments"
-import { diffAllowances } from "./diffAllowance"
+import { diffAllowances } from "./allowances"
 
 const key1 = hexlify(zeroPadValue("0x01", 32)) as `0x${string}`
 const key2 = hexlify(zeroPadValue("0x02", 32)) as `0x${string}`
